+++ test/model/session.js
@@ -7,13 +7,19 @@
 class Session {
   constructor(options= {}) {
     if (typeof options === 'string') {
-      this.name = options;
+      this.user = {
+        username: options,
+        id: 1
+      };
     } else {
-      this.name = options.name === undefined ? 'John' : options.name;
-      this.id = options.id === undefined ? 1 : options.id;
+      this.user = {
+        username:  options.name === undefined ? 'John' : options.name,
+        id: options.id === undefined ? 1 : options.id,
+      }
       this.reason = options.reason === undefined ? undefined : options.reason;
     }
   }
 }
+
 
 module.exports = Session;