+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Mongoose Undo
 ## revisions
 
+## 2021-06-04 (0.2.0)
+- change: session has now an user object with a username variable
+
 ## 2020-05-06
-- added Model.ession
+- added Model.session
 - session.setInfo(Model) to create a custom session definition
 
 ## 2020-04-27